requirements.txt
beautifulsoup4
flask==1.1.4
flask-mysql==1.5.2
markupsafe==2.0.1
mysql-connector-python
numpy
pandas
requests
scikit-metrics
timeago==1.0.15
selenium



